User commit

+++ src/main/java/com/example/demo/repository/StudentRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.Studententity;
 
-public interface StudentRepo extends JpaRepository<Studententity, int> {
+public interface StudentRepo extends JpaRepository<Studententity, Long> {
 }
